fix deploy

+++ src/main/java/com/example/backendbase/manager/repo/native_repo/ServiceNativeRepo.java
@@ -45,7 +45,7 @@ public class ServiceNativeRepo {
             handOverServiceDTO.setConntractId(((BigInteger) record[3]).longValue());
             handOverServiceDTO.setServiceShowName((String) record[4]);
             handOverServiceDTO.setServiceType(((BigInteger) record[5]).longValue());
-            handOverServiceDTO.setId(((BigInteger) record[6]).longValue());
+            handOverServiceDTO.setId(((Integer) record[6]).longValue());
             result.add(handOverServiceDTO);
         });
         return result;
